api-server/requirements.txt
# Dépendances Python pour l'API Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
minio==7.2.0
redis==5.0.1
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2